refactor: move match templates and config to events app
- Move `MatchTemplate`, `MatchTemplateItem`, and `EventMatchConfiguration` from `matches` app to `events` app.
- Rename models to `EventMatchTemplate` and `EventMatchTemplateItem` for clarity.
- Transfer associated Service logic, Serializers, Views, and URL registration to `events`.
- Update Admin registration to reflect the model relocation.
- Update database constraints in `events` models to use dynamic naming patterns.
- Add `setup_prod` command to Makefile.
